Delete Person: Allows administrators to delete people from the person's edit screen.
NOTE: for some reason unknown to me, the "after_destroy" call back in index_observer.rb was not functioning properly (i.e., works were not being reindexed when a person was deleted). I changed it to "before_destroy" and it works. If this turns out to have adverse effects on other models, then we'll have to come up with a better solution.