Groups: Show:	*--- requires solr refresh ---*
In response to Issue 76:

Adds group_ids to Person.to_solr_data as a serialized array.
Group: Show pages now check to see if the person is actually 
a part of the group, rather than a co-author on a work with 
a member of the group, as was happening before.